springSecurity +redis+jwt搭建

+++ commons/src/main/java/cn/rmonkey/commons/entity/BaseEntity.java
@@ -15,6 +15,6 @@ import java.io.Serializable;
 @Setter
 public abstract class BaseEntity implements Serializable {
     @TableId
-    protected Integer id;
+    protected Long id;
 
 }
